project-healthcare/requirements.txt
tensorflow
googlemaps
geopy
langchain
uvicorn
langchain-openai
langchain-community
langchain-text-splitters
langchain-core
faiss-cpu
fastapi
streamlit
langgraph
python-dotenv
pydantic
langchain-google-genai
pandas
numpy
